# perf(webidl/ByteString): 3x faster ASCII check