--- a/ext/webidl/00_webidl.js
+++ b/ext/webidl/00_webidl.js
@@ -375,10 +375,10 @@
   };
 
   // deno-lint-ignore no-control-regex
-  const IS_ASCII = /^[\x00-\xFF]*$/;
+  const IS_BYTESTRING = /^[\x00-\xFF]*$/;
   converters.ByteString = (V, opts) => {
     const x = converters.DOMString(V, opts);
-    if (!RegExpPrototypeTest(IS_ASCII, x)) {
+    if (!RegExpPrototypeTest(IS_BYTESTRING, x)) {
       throw makeException(TypeError, "is not a valid ByteString", opts);
     }
     return x;